dom/base/test/test_meta_viewport_auto_size_by_device_width.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>device-width enables autoSize</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="width=device-width">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>width=device-width</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function device_width() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      ok(info.autoSize, "device-width should enable autoSize");
    });
  </script>
</body>
</html>
